+++ bin/config.ml
@@ -5,28 +5,42 @@ let invalid_config_key key =
 
 let show_val = function None -> "<unset>" | Some x -> x
 
-let log_val string_opt =
+let log_val s =
+  Logs.app (fun l -> l "%s" s);
+  Ok ()
+
+let log_val_opt string_opt =
   Logs.app (fun l -> l "%s" (show_val string_opt));
   Ok ()
 
+let no_config_message =
+  "You don't have a dune-release config file yet. You can create one by \
+   running `dune-release config create` or simply wait for dune-release to \
+   prompt you when it will actually need it."
+
 let show key =
   let open Rresult.R.Infix in
-  Config.load () >>= fun config ->
-  match key with
+  Config.load () >>= function
   | None ->
-      let pretty_fields = Config.pretty_fields config in
-      StdLabels.List.iter pretty_fields ~f:(fun (key, value) ->
-          Logs.app (fun l -> l "%s: %s" key (show_val value)));
+      App_log.status (fun l -> l "%s" no_config_message);
       Ok ()
-  | Some "user" ->
-      Logs.warn (fun l -> l "%s" Deprecate.Config_user.config_field_use);
-      log_val config.user
-  | Some "remote" -> log_val config.remote
-  | Some "local" -> log_val (Stdext.Option.map ~f:Fpath.to_string config.local)
-  | Some "keep-v" -> log_val (Stdext.Option.map ~f:string_of_bool config.keep_v)
-  | Some "auto-open" ->
-      log_val (Stdext.Option.map ~f:string_of_bool config.auto_open)
-  | Some key -> invalid_config_key key
+  | Some config -> (
+      match key with
+      | None ->
+          let pretty_fields = Config.pretty_fields config in
+          StdLabels.List.iter pretty_fields ~f:(fun (key, value) ->
+              Logs.app (fun l -> l "%s: %s" key (show_val value)));
+          Ok ()
+      | Some "user" ->
+          Logs.warn (fun l -> l "%s" Deprecate.Config_user.config_field_use);
+          log_val_opt config.user
+      | Some "remote" -> log_val config.remote
+      | Some "local" -> log_val (Fpath.to_string config.local)
+      | Some "keep-v" ->
+          log_val_opt (Stdext.Option.map ~f:string_of_bool config.keep_v)
+      | Some "auto-open" ->
+          log_val_opt (Stdext.Option.map ~f:string_of_bool config.auto_open)
+      | Some key -> invalid_config_key key)
 
 let to_bool ~field value =
   match String.lowercase_ascii value with
@@ -36,22 +50,38 @@ let to_bool ~field value =
 
 let set key value =
   let open Rresult.R.Infix in
-  Config.load () >>= fun config ->
-  let updated =
-    match key with
-    | "user" ->
-        App_log.unhappy (fun l -> l "%s" Deprecate.Config_user.config_field_use);
-        Ok { config with user = Some value }
-    | "remote" -> Ok { config with remote = Some value }
-    | "local" ->
-        Fpath.of_string value >>| fun v -> { config with local = Some v }
-    | "keep-v" ->
-        to_bool ~field:key value >>| fun v -> { config with keep_v = Some v }
-    | "auto-open" ->
-        to_bool ~field:key value >>| fun v -> { config with auto_open = Some v }
-    | _ -> invalid_config_key key
-  in
-  updated >>= Config.save >>= fun () -> Ok ()
+  Config.load () >>= function
+  | None -> Rresult.R.error_msgf "%s" no_config_message
+  | Some config ->
+      let updated =
+        match key with
+        | "user" ->
+            App_log.unhappy (fun l ->
+                l "%s" Deprecate.Config_user.config_field_use);
+            Ok { config with user = Some value }
+        | "remote" -> Ok { config with remote = value }
+        | "local" ->
+            Fpath.of_string value >>| fun v -> { config with local = v }
+        | "keep-v" ->
+            to_bool ~field:key value >>| fun v ->
+            { config with keep_v = Some v }
+        | "auto-open" ->
+            to_bool ~field:key value >>| fun v ->
+            { config with auto_open = Some v }
+        | _ -> invalid_config_key key
+      in
+      updated >>= Config.save >>= fun () -> Ok ()
+
+let create () =
+  let open Rresult.R.Infix in
+  Config.load () >>= function
+  | None -> Config.create ()
+  | Some _ ->
+      App_log.status (fun l ->
+          l
+            "You already have a dune-release configuration file. Use \
+             `dune-release config set` to modify it.");
+      Ok ()
 
 let default_usage ?raw () =
   let cmd = "dune-release config" in
@@ -72,6 +102,10 @@ let set_usage ?raw () =
   | Some () -> Printf.sprintf "%s %s %s" cmd key value
   | None -> Printf.sprintf "$(b,%s) $(i,%s) $(i,%s)" cmd key value
 
+let create_usage ?raw () =
+  let cmd = "dune-release config create" in
+  match raw with Some () -> cmd | None -> Printf.sprintf "$(b,%s)" cmd
+
 let invalid_usage () =
   Rresult.R.error_msgf
     "Invalid dune-release config invocation. Usage:\n%s\n%s\n%s"
@@ -83,6 +117,7 @@ let run action key_opt value_opt =
      match (action, key_opt, value_opt) with
      | "show", key, None -> show key
      | "set", Some key, Some value -> set key value
+     | "create", None, None -> create ()
      | _ -> invalid_usage ()
    in
    res >>= fun () -> Ok 0)
@@ -95,6 +130,7 @@ let man =
     `P (default_usage ());
     `P (show_usage ());
     `P (set_usage ());
+    `P (create_usage ());
     `S "GLOBAL CONFIGURATION FIELDS";
     `P
       "Here are the existing fields of dune-release's global config file. Only \